Added i3 support + ConfigFileEditor
# ConfigFileEditor
Instead of having a configuration file where we define the colors, and another generated by hapycolor, we can use only one file, and define the colors to replace with some a macro like this:
```
? @hapycolor.color("foreground")
color_1 #010203
? @hapycolor.color("random")
color_2 #123456
```
where `?` is the character that defines the beginning of a comment. Hapycolor should be able to identify the colors defined in the following line, and replace them with the ones contained in the palette.

# i3 Support
The ConfigFileEditor has been added to support custom colors in i3's configuration file, and replaces the wallpaper with the selected one, if the latter feature is enabled.